<fix> transaciton value precision - use decimal instead of double to fix multiprecision sum of transacions

+++ src/Draftor.Application/Interfaces/IPersonService.cs
@@ -12,7 +12,7 @@ public interface IPersonService
 
     Task AddTransactionAsync(TransactionVM transaction);
 
-    Task AddTransactionBulk(double ammount, string title, string description, IEnumerable<int> peopleChecked);
+    Task AddTransactionBulk(decimal ammount, string title, string description, IEnumerable<int> peopleChecked);
 
     Task DeletePersonAsync(int id);
